+++ packages/subgraph/src/dao/dao_v1_0_0.ts
@@ -2,7 +2,7 @@ import {
   Dao,
   Permission,
   StandardCallback,
-  TransactionActionsProposal,
+  TransactionActions as TransactionActionsProposal,
 } from '../../generated/schema';
 import {
   MetadataSet,