+++ packages/subgraph/src/dao/dao_v1_3_0.ts
@@ -1,48 +1,51 @@
-import {TransactionActionsProposal} from '../../generated/schema';
+import {TransactionActions} from '../../generated/schema';
 import {
   Executed,
   ExecutedActionsStruct,
 } from '../../generated/templates/DaoTemplateV1_3_0/DAO';
 import {handleAction} from './utils';
+import {generateDaoEntityId} from '@aragon/osx-commons-subgraph';
 import {
-  generateDaoEntityId,
-  generateProposalEntityId,
-  generateTransactionActionsProposalEntityId,
-} from '@aragon/osx-commons-subgraph';
-import {BigInt} from '@graphprotocol/graph-ts';
+  generateTransactionActionsDeterministicId,
+  generateTransactionActionsEntityId,
+} from './ids';
+import {Bytes} from '@graphprotocol/graph-ts';
 
 export function handleExecuted(event: Executed): void {
-  let proposalEntityId = generateProposalEntityId(
-    event.params.actor,
-    BigInt.fromByteArray(event.params.callId)
+  let transactionActionsEntityId = generateTransactionActionsEntityId(
+    event.address.toHexString(),
+    event.params.actor.toHexString(),
+    event.params.callId.toString(),
+    event.transaction.hash,
+    event.transactionLogIndex
   );
 
-  let transactionActionsProposalEntityId =
-    generateTransactionActionsProposalEntityId(
-      proposalEntityId,
-      event.transaction.hash,
-      event.transactionLogIndex
-    );
-  let proposal = new TransactionActionsProposal(
-    transactionActionsProposalEntityId
+  let deterministicId = generateTransactionActionsDeterministicId(
+    event.address.toHexString(),
+    event.params.actor.toHexString(),
+    event.params.callId.toString()
   );
-  proposal.dao = generateDaoEntityId(event.address);
-  proposal.createdAt = event.block.timestamp;
-  proposal.endDate = event.block.timestamp;
-  proposal.startDate = event.block.timestamp;
-  proposal.creator = event.params.actor;
-  proposal.executionTxHash = event.transaction.hash;
-  proposal.allowFailureMap = event.params.allowFailureMap;
-  proposal.executed = true;
-  proposal.failureMap = event.params.failureMap;
-  proposal.save();
+
+  let transactionActions = new TransactionActions(transactionActionsEntityId);
+
+  transactionActions.dao = generateDaoEntityId(event.address);
+  transactionActions.deterministicId = Bytes.fromHexString(deterministicId);
+  transactionActions.createdAt = event.block.timestamp;
+  transactionActions.endDate = event.block.timestamp;
+  transactionActions.startDate = event.block.timestamp;
+  transactionActions.creator = event.params.actor;
+  transactionActions.executionTxHash = event.transaction.hash;
+  transactionActions.allowFailureMap = event.params.allowFailureMap;
+  transactionActions.executed = true;
+  transactionActions.failureMap = event.params.failureMap;
+  transactionActions.save();
 
   let actions = event.params.actions;
 
   for (let index = 0; index < actions.length; index++) {
     handleAction<ExecutedActionsStruct, Executed>(
       actions[index],
-      transactionActionsProposalEntityId,
+      transactionActionsEntityId,
       index,
       event
     );