+++ packages/subgraph/src/dao/utils.ts
@@ -1,9 +1,9 @@
 import {
-  Action,
   AddresslistVotingProposal,
   AdminProposal,
   MultisigProposal,
   TokenVotingProposal,
+  TransactionAction,
 } from '../../generated/schema';
 import {
   Executed,
@@ -24,7 +24,8 @@ import {handleERC721Action} from '../utils/tokens/erc721';
 import {handleERC1155Action} from '../utils/tokens/erc1155';
 import {handleNativeAction} from '../utils/tokens/eth';
 import {generateDaoEntityId} from '@aragon/osx-commons-subgraph';
-import {BigInt} from '@graphprotocol/graph-ts';
+import {BigInt, Bytes} from '@graphprotocol/graph-ts';
+import {generateActionEntityId, generateDeterministicActionId} from './ids';
 
 // AssemblyScript struggles having multiple return types. Due to this,
 // The below seems most effective way.
@@ -84,34 +85,58 @@ export function handleAction<
     return;
   }
 
-  handleTokenTransfers(action, proposalId, index, event);
+  checkForAndHandleTokenTransfers(action, proposalId, index, event);
 }
 
 function getOrCreateActionEntity<
   T extends ExecutedActionsStruct,
   R extends Executed
->(action: T, proposalId: string, index: i32, event: R): Action {
-  const actionId = [proposalId, index.toString()].join('_');
-  let entity = Action.load(actionId);
+>(
+  action: T,
+  transactionActionId: string,
+  index: i32,
+  event: R
+): TransactionAction {
+  const deterministicActionId = generateDeterministicActionId(
+    event.address.toHexString(),
+    event.params.actor.toHexString(),
+    event.params.callId.toString(),
+    index
+  );
+  const actionId = generateActionEntityId(
+    event.address.toHexString(),
+    event.params.actor.toHexString(),
+    event.params.callId.toString(),
+    index,
+    event.transaction.hash,
+    event.transactionLogIndex
+  );
 
-  // In case the execute on the dao is called by the address
-  // That we don't currently index for the actions in the subgraph,
-  // we fallback and still create an action.
-  // NOTE that it's important to generate action id differently to not allow
+  /// we shouldn't need to load the entity as it should be created each time
+  let entity = TransactionAction.load(actionId);
 
   if (!entity) {
-    entity = new Action(actionId);
+    entity = new TransactionAction(actionId);
+    entity.deterministicId = Bytes.fromHexString(deterministicActionId);
     entity.to = action.to;
     entity.value = action.value;
     entity.data = action.data;
-    entity.proposal = proposalId;
+    entity.transactionActions = transactionActionId;
     entity.dao = generateDaoEntityId(event.address);
   }
 
   return entity;
 }
 
-function handleTokenTransfers<
+/**
+ * Determines if the action is an ERC20, ERC721 or ERC1155 transfer and calls the appropriate handler if so.
+ * Does nothing if the action is not a recognised token transfer.
+ * @param action the action to validate
+ * @param proposalId TODO: change this
+ * @param actionIndex the index number of the action inside the executed batch
+ * @param event the Executed event emitting the event
+ */
+function checkForAndHandleTokenTransfers<
   T extends ExecutedActionsStruct,
   R extends Executed
 >(action: T, proposalId: string, actionIndex: i32, event: R): void {