+++ packages/subgraph/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UPCOMING]
 
+## Split
+
+- Added deterministicID as field to `TransactionActions` and `TransactionAction`
+- Temp entity `TransactionAction` created while plugins (with actions) still live inside OSx. Includes `deterministicId` and swaps `proposal` foreign key in place of `transactionAction`
+- The `deterministicId` and `Id` fields now include the `daoAddress` inside executeActions
+- Renamed the `handleTokenTransfer` inside the `handleAction` to better reflect the fact that it checks _all_ actions.
+
 ## 1.4.1
 
 ## 1.5.0